packages/core: use cleaner tree of index files in api + remove default exports
