+++ packages/core/src/api/apis/ApiProvider.test.tsx
@@ -15,9 +15,9 @@
  */
 
 import React from 'react';
-import ApiProvider, { useApi, withApis } from './ApiProvider';
-import ApiRef from './ApiRef';
-import ApiRegistry from './ApiRegistry';
+import { ApiProvider, useApi, withApis } from './ApiProvider';
+import { ApiRef } from './ApiRef';
+import { ApiRegistry } from './ApiRegistry';
 import { render } from '@testing-library/react';
 import { withLogCollector } from '@backstage/test-utils-core';
