+++ packages/core/src/api/apis/ApiRef.test.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import ApiRef from './ApiRef';
+import { ApiRef } from './ApiRef';
 
 describe('ApiRef', () => {
   it('should be created', () => {
