+++ packages/core/src/api/apis/definitions/featureFlags.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import ApiRef from '../ApiRef';
+import { ApiRef } from '../ApiRef';
 import {
   UserFlags,
   FeatureFlagsRegistry,
